auth-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-keycloak==3.7.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
flask-limiter==3.5.0
flask-smorest==0.42.3
marshmallow>=3.20.1